feat: move and rename NotePopover to WaklthroughPanel
