+++ packages/diagram/src/navigationpanel/walkthrough/WalkthroughPanel.css.ts
